# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Collectarr",
    "description": "Collectarr metadata and library backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/catalog/summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Catalog Summary",
        "operationId": "catalog_summary_api_v1_admin_catalog_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCatalogSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/catalog/normalized-metadata-drift": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Catalog Normalized Metadata Drift",
        "operationId": "catalog_normalized_metadata_drift_api_v1_admin_catalog_normalized_metadata_drift_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Sample Limit"
            }
          },
          {
            "name": "scan_limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100000,
                  "minimum": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scan Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminNormalizedMetadataDriftReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/search/status": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Search Status",
        "operationId": "search_status_api_v1_admin_search_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSearchStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/search/reindex": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Search Reindex",
        "operationId": "search_reindex_api_v1_admin_search_reindex_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSearchReindexResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/search/history": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Search History",
        "operationId": "search_history_api_v1_admin_search_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminSearchHistoryEntry"
                  },
                  "type": "array",
                  "title": "Response Search History Api V1 Admin Search History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/audit/logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Audit Logs",
        "operationId": "audit_logs_api_v1_admin_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminAuditLogResponse"
                  },
                  "title": "Response Audit Logs Api V1 Admin Audit Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/duplicates": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Duplicate Candidates",
        "operationId": "duplicate_candidates_api_v1_admin_duplicates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminDuplicateCandidateResponse"
                  },
                  "title": "Response Duplicate Candidates Api V1 Admin Duplicates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/duplicates/summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Duplicate Queue Summary",
        "operationId": "duplicate_queue_summary_api_v1_admin_duplicates_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDuplicateQueueSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/duplicates/reviews": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Duplicate Review History",
        "operationId": "duplicate_review_history_api_v1_admin_duplicates_reviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminDuplicateReviewEntryResponse"
                  },
                  "title": "Response Duplicate Review History Api V1 Admin Duplicates Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/duplicates/ignore": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Ignore Duplicate Candidate",
        "operationId": "ignore_duplicate_candidate_api_v1_admin_duplicates_ignore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDuplicateIgnoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDuplicateActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/duplicates/merge": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Merge Duplicate Candidate",
        "operationId": "merge_duplicate_candidate_api_v1_admin_duplicates_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDuplicateMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDuplicateActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/duplicates/review": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Review Duplicate Candidate",
        "operationId": "review_duplicate_candidate_api_v1_admin_duplicates_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDuplicateReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDuplicateActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_v1_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__schemas__admin__UserResponse"
                  },
                  "type": "array",
                  "title": "Response List Users Api V1 Admin Users Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get User",
        "operationId": "get_user_api_v1_admin_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__admin__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update User",
        "operationId": "update_user_api_v1_admin_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__admin__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/image-cache/stats": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Image Cache Stats",
        "operationId": "image_cache_stats_api_v1_admin_image_cache_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageCacheStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/image-cache/purge": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Image Cache Purge",
        "operationId": "image_cache_purge_api_v1_admin_image_cache_purge_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageCachePurgeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AdminAuditLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "actor_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Email"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "details_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action",
          "entity_type",
          "details_json",
          "created_at"
        ],
        "title": "AdminAuditLogResponse"
      },
      "AdminCatalogSummaryResponse": {
        "properties": {
          "items": {
            "type": "integer",
            "title": "Items"
          },
          "items_by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Items By Kind"
          },
          "image_assets": {
            "type": "integer",
            "title": "Image Assets"
          },
          "image_cache_entries": {
            "type": "integer",
            "title": "Image Cache Entries"
          },
          "missing_cover_items": {
            "type": "integer",
            "title": "Missing Cover Items"
          },
          "duplicate_candidate_groups": {
            "type": "integer",
            "title": "Duplicate Candidate Groups"
          }
        },
        "type": "object",
        "required": [
          "items",
          "image_assets",
          "image_cache_entries",
          "missing_cover_items",
          "duplicate_candidate_groups"
        ],
        "title": "AdminCatalogSummaryResponse"
      },
...
      "AdminDuplicateCandidateResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "item_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Number"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "item_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Item Ids"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": "same title and item number"
          },
          "has_cover_conflicts": {
            "type": "boolean",
            "title": "Has Cover Conflicts",
            "default": false
          },
          "duplicate_score": {
            "type": "integer",
            "title": "Duplicate Score",
            "default": 0
          },
          "recommended_target_item_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Target Item Id"
          },
          "confidence_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Confidence Factors"
          },
          "merge_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Merge Warnings"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "title",
          "item_number",
          "count",
          "item_ids"
        ],
        "title": "AdminDuplicateCandidateResponse"
      },
...
      "AdminNormalizedMetadataDriftReportResponse": {
        "properties": {
          "expected_schema_version": {
            "type": "integer",
            "title": "Expected Schema Version"
          },
          "scan_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Limit"
          },
          "scan_limited": {
            "type": "boolean",
            "title": "Scan Limited",
            "default": false
          },
          "scanned_entities": {
            "type": "integer",
            "title": "Scanned Entities",
            "default": 0
          },
          "entities_with_normalized": {
            "type": "integer",
            "title": "Entities With Normalized",
            "default": 0
          },
          "drifted_entities": {
            "type": "integer",
            "title": "Drifted Entities",
            "default": 0
          },
          "typed_scanned_items": {
            "type": "integer",
            "title": "Typed Scanned Items",
            "default": 0
          },
          "typed_drifted_items": {
            "type": "integer",
            "title": "Typed Drifted Items",
            "default": 0
          },
          "schema_issue_count": {
            "type": "integer",
            "title": "Schema Issue Count",
            "default": 0
          },
          "blocking_issue_count": {
            "type": "integer",
            "title": "Blocking Issue Count",
            "default": 0
          },
          "release_gate_ok": {
            "type": "boolean",
            "title": "Release Gate Ok",
            "default": true
          },
          "issue_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Issue Counts"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/AdminNormalizedMetadataDriftSample"
            },
            "type": "array",
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "expected_schema_version"
        ],
        "title": "AdminNormalizedMetadataDriftReportResponse"
      },
...
      "AdminSearchHistoryEntry": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "index_name": {
            "type": "string",
            "title": "Index Name"
          },
          "indexed_documents": {
            "type": "integer",
            "title": "Indexed Documents"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "ok",
          "index_name",
          "indexed_documents"
        ],
        "title": "AdminSearchHistoryEntry"
      },
      "AdminSearchReindexResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "index_name": {
            "type": "string",
            "title": "Index Name"
          },
          "indexed_documents": {
            "type": "integer",
            "title": "Indexed Documents"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "index_name",
          "indexed_documents"
        ],
        "title": "AdminSearchReindexResponse"
      },
      "AdminSearchStatusResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "index_name": {
            "type": "string",
            "title": "Index Name"
          },
          "document_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Count"
          },
          "is_empty": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Empty"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "index_name"
        ],
        "title": "AdminSearchStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}